A bunch of small mods to get it working on the mingw32 compiler

+++ src/game.cpp
@@ -12,7 +12,7 @@ using namespace std;
 static void handleEvents(GameState *st);
 
 static bool running;
-static int fps;
+static float fps;
 
 
 /**
